+++ packages/sampler/stories/default/Heading.js
@@ -1,35 +1,39 @@
-import Divider, {DividerBase} from '@enact/moonstone/Divider';
+import Heading, {HeadingBase} from '@enact/moonstone/Heading';
+import UiHeading from '@enact/ui/Heading';
 import BodyText from '@enact/moonstone/BodyText';
 import React from 'react';
 import {storiesOf} from '@storybook/react';
 import {withInfo} from '@storybook/addon-info';
 
-import {select, text} from '../../src/enact-knobs';
+import {boolean, select, text} from '../../src/enact-knobs';
 import {mergeComponentMetadata} from '../../src/utils';
 
-Divider.displayName = 'Divider';
-const Config = mergeComponentMetadata('Divider', DividerBase, Divider);
+Heading.displayName = 'Heading';
+const Config = mergeComponentMetadata('Heading', UiHeading, HeadingBase, Heading);
 
 // Set up some defaults for info and knobs
 const prop = {
 	casing: ['', 'preserve', 'sentence', 'word', 'upper'],
 	marqueeOn: ['', 'hover', 'render'],
-	spacing: ['', 'normal', 'small', 'medium', 'large', 'none']
+	size: ['', 'large', 'medium', 'small'],
+	spacing: ['', 'auto', 'large', 'medium', 'small', 'none']
 };
 
 storiesOf('Moonstone', module)
 	.add(
-		'Divider',
+		'Heading',
 		withInfo({
-			text: 'Basic usage of Divider'
+			text: 'A component for initiating a section of content.'
 		})(() => (<React.Fragment>
-			<Divider
+			<Heading
 				casing={select('casing', prop.casing, Config)}
 				marqueeOn={select('marqueeOn', prop.marqueeOn, Config)}
+				showLine={boolean('showLine', Config)}
+				size={select('size', prop.size, Config)}
 				spacing={select('spacing', prop.spacing, Config)}
 			>
-				{text('children', Config, 'divider text')}
-			</Divider>
+				{text('children', Config, 'Heading text')}
+			</Heading>
 			<BodyText style={{marginTop: 0}}>
 				Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed aliquam dapibus imperdiet. Morbi diam ex, vulputate eget luctus eu, gravida at ligula. Sed tristique eros sit amet iaculis varius. Phasellus rutrum augue id nulla consectetur, a vulputate velit dictum. Vestibulum ultrices tellus ac cursus condimentum. Aliquam sit amet consectetur nulla, viverra bibendum metus.
 			</BodyText>