+++ packages/ui/CHANGELOG.md
@@ -9,6 +9,7 @@ The following is a curated list of changes in the Enact ui module, newest change
 - `ui/Measurable` HOC and Hook for quick and convenient measuring of simple components
 - `ui/Button` public class `.hasIcon` which is present on the root node only when an icon has been provided
 - `ui/Scroller`, `ui/VirtualList.VirtualGridList`, and `ui/VirtualList.VirtualList` prop `noScrollByWheel` for preventing scroll by wheel
+- `ui/Heading` component
 
 ### Fixed